add bldg alert data into wow_indicators & update API (#1033)
Adds to the new wow_indicators table we created for area alerts to incorporate all the indicators for building alert emails. With this change now all email indicator data is pre-calculated for the necessary date ranges. This allows for a new building alert data API that takes a list of one or more BBLs and returns all the indicator data for the week and month values, so now there is only a single API request per building alert email rather than a weekly and monthly request for each building and each requires a lot of calculation on the DB, which does not scale well and has made the email sending process extremely slow and taxing for the DB.

Companion PR to update the building alert email sending process/template: JustFixNYC/auth-provider#101