add back temporarily removed gce eligibility tables (#1038)
In #1007 we temporarily had to remove the GCE eligibility and map tables because they rely on pluto_latest_districts which was a one-off table we created directly on the DB outside of our usual nycdb process and that broke tests. We have since updated nycdb to add this dataset with a new pattern for datasets without source data and with dependencies on other nycdb datasets in nycdb/nycdb#386. So now we can add these tables back to the goodcause job to run on k8s.

Companion PR: JustFixNYC/nycdb-k8s-loader#216