upgrade to node 16 (#1051)
My local docker was failing to build because the source files for node v12 are no longer available for download. To get this working I needed to upgrade to 16. This PR changes both the docker setup (which is only used for local dev) and the versio set in package.json and I've tested it out by updating the engine version on netfliy for the demo site only, and those builds are all passing (the prod ones fail since its still on 12.x

The only package I ran into problems with for this change was one of our own NPM packages that we use for common strings, but it was only used in an old covid banner that isn't active anymore so I've removed all that code and the dependency.

16 is still past end-of-life so we should really get it up further, but I'll follow up with a new PRs soon hopefully to keep bumping it up.

So if this all looks good we should be able to make the same change on netlify for the prod site and push main to prod. I think there's a slight risk of the build breaking while trying to get that timing right, so probably best to do at a low traffic time, so I can try to do this late tonight